+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Created on Dec 20, 2023
 - Initial Create: Initial created on December 14, 2023 -> **0.00**
 - Initial Development: Development started on Decemeber 20, 2023 -> **0.1**
 - Initial Development: Authentication processes work on Decemeber 21, 2023 -> **0.2**
+- Initial Development: Account manager is completed on Decemeber 21, 2023 -> **0.22**